+++ src/actions/index.ts
@@ -9,3 +9,4 @@ export * from './validate';
 export * from './store';
 export * from './login';
 export * from './logout';
+export * from './generate';