+++ src/propagation_server/index.ts
@@ -1,3 +1,6 @@
 import { startPreviewServer } from './server';
+import tasks from "./tasks";
 
+
+tasks.start();
 startPreviewServer();