+++ src/yargs/setupCommands.ts
@@ -13,6 +13,7 @@ import {
   loginCommand, 
   logoutCommand,
   serverCommand,
+  generateCredsCommand
 } from "./commands";
 
 // Configure and run Yargs
@@ -34,6 +35,7 @@ export async function setupCommands() {
   loginCommand(parser);
   logoutCommand(parser);
   serverCommand(parser);
+  generateCredsCommand(parser);
 
   // Set default command and help
   parser